refactor: move Preferences into Menu folder
